requirements.txt
python-amazon-paapi>=5.0.0
twilio>=8.0.0
python-dotenv>=1.0.0
